ingestor/pyproject.toml
[project]
name = "binance-ingestor"
version = "0.1.0"
description = "Standalone Binance data ingestor for duckport-rs"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "pyarrow>=15.0",
    "aiohttp>=3.9",
    "websockets>=13.0",
    "python-dotenv>=1.0",
    "colorama>=0.4",
    "pytz",
    "duckdb>=1.0",
    "lxml>=5.0",
    "aiofiles>=24.0",
    "tqdm>=4.60",
    "joblib>=1.3",
    "python-dateutil>=2.8",
    "numpy>=1.26",
]

[tool.setuptools.packages.find]
include = ["binance_ingestor*"]

[project.scripts]
binance-ingestor = "binance_ingestor.main:main"
loadhist = "binance_ingestor.loadhist:main"
